add:cloudflare turnstileのために機能分割化
