+++ frontend-flare/src/app/login/page.tsx
@@ -1,62 +1,12 @@
-"use client";
-
-import { createServerComponentClient } from "@/lib/supabase/utils";
-import { headers } from "next/headers";
-import { redirect } from "next/navigation";
-import { Button } from "../components/ui/Button";
-import { useState } from "react";
-import { Turnstile } from "@marsidev/react-turnstile";
+import { LoginForm } from "./LoginForm";
 
 type LoginPageProps = {
-  searchParams?: { message?: string };
+  searchParams?: Promise<{ message?: string }>;
 };
 
-const GoogleIcon = () => (
-  <svg className="w-5 h-5 mr-3" viewBox="0 0 24 24">
-    <path
-      d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-      fill="#4285F4"
-    />
-    <path
-      d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-      fill="#34A853"
-    />
-    <path
-      d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-      fill="#FBBC05"
-    />
-    <path
-      d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-      fill="#EA4335"
-    />
-  </svg>
-);
-
-// Server Action for Google Sign-In
-const signInWithGoogle = async () => {
-  "use server";
-
-  const origin = (await headers()).get("origin");
-  const supabase = await createServerComponentClient();
-
-  const { data, error } = await supabase.auth.signInWithOAuth({
-    provider: "google",
-    options: {
-      redirectTo: `${origin}/auth/callback`,
-    },
-  });
-
-  if (error) {
-    console.error(error);
-    return redirect("/login?message=Could not authenticate with Google");
-  }
-
-  return redirect(data.url);
-};
-
-export default function Login({ searchParams }: LoginPageProps) {
-  const [token, setToken] = useState<string | null>(null);
-  const message = searchParams?.message;
+export default async function Login({ searchParams }: LoginPageProps) {
+  const resolvedSearchParams = await searchParams;
+  const message = resolvedSearchParams?.message;
 
   return (
     <div className="min-h-screen w-full flex items-center justify-center bg-[#0a0a0a] relative overflow-hidden">
@@ -77,28 +27,7 @@ export default function Login({ searchParams }: LoginPageProps) {
           </p>
         </div>
 
-        <form action={signInWithGoogle} className="space-y-6">
-          <Button
-            variant="default"
-            size="lg"
-            className="w-full h-12 bg-white hover:bg-gray-100 text-gray-900 font-medium border border-gray-200 transition-all hover:shadow-lg hover:scale-[1.02] disabled:opacity-50 disabled:cursor-not-allowed disabled:scale-100"
-            disabled={!token}
-          >
-            <GoogleIcon />
-            Continue with Google
-          </Button>
-
-          <Turnstile
-            siteKey={process.env.NEXT_PUBLIC_TURNSTILE_SITE_KEY!}
-            onSuccess={setToken}
-            onExpire={() => setToken(null)}
-            onError={() => setToken(null)}
-            options={{
-              theme: "dark",
-            }}
-            className="mx-auto"
-          />
-        </form>
+        <LoginForm />
 
         {message && (
           <div className="p-4 bg-red-500/10 border border-red-500/20 rounded-xl">
@@ -115,3 +44,4 @@ export default function Login({ searchParams }: LoginPageProps) {
     </div>
   );
 }
+
